--- a/babel.config.js
+++ b/babel.config.js
@@ -1,13 +1,10 @@
 module.exports = {
   presets: [['@babel/env', { loose: true }], '@babel/preset-react'],
-  // plugins: ['@babel/transform-runtime'],
+  plugins: ['@babel/plugin-syntax-import-assertions'],
   env: {
     esm: {
-      presets: [
-        ['@babel/env', { loose: true, modules: false }],
-        '@babel/preset-react',
-      ],
+      presets: [['@babel/env', { loose: true, modules: false }], '@babel/preset-react'],
       // plugins: [['@babel/transform-runtime', { useESModules: true }]],
     },
   },
-}
+};