--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tradersclub/styled-system",
-  "version": "6.0.0--beta5",
+  "version": "6.0.0--beta7",
   "typings": "index.d.ts",
   "main": "dist/index.cjs.js",
   "module": "dist/index.esm.js",