--- a/package.json
+++ b/package.json
@@ -33,7 +33,7 @@
     "prettier": "2.3.2",
     "react": "18.2.0",
     "react-test-renderer": "16.13.1",
-    "styled-components": "6.0.4"
+    "styled-components": "6.0.5"
   },
   "peerDependencies": {
     "styled-components": ">=6"