# V 6.0.0
- Styled components atualizado para 6.0
- Refatoração da lógica de processamento/seleção dinâmica de props do styled-components


## Resultados preliminares TCWeb:


[Profiles.zip](https://github.com/styled-system/styled-system/files/12152877/Profiles.zip)

### Antes:
![Captura de Tela 2023-07-24 às 19 07 10](https://github.com/styled-system/styled-system/assets/71399856/76fd5e69-54aa-4491-923e-63b184b635ad)

### Depois:
![Captura de Tela 2023-07-24 às 19 07 24](https://github.com/styled-system/styled-system/assets/71399856/a4040b1e-7988-4d59-be4a-77278afa0318)


### Análise:

Avaliando os profiles do Devtools é possível observar que o primeiro render diminuiu em média 66% (2/3), de ~667ms para ~219ms;  Componentes tiveram uma melhoria de até 80% no tempo para serem renderizados, sendo que com a melhoria nenhum componente levou mais de 20ms para renderizar, onde anteriormente a maioria levava em média 40ms atingindo 80ms+ nos seus picos.

O TTI (time to interactive, quando a página está pronta para receber interações do usuário) caiu de ~5,5s para ~4,4s, sendo que o FCP(First Contentful Paint) passou a atingir a marca de aceitável nos padrões da Google na marca de ~1,72s se comparado com o tempo de ~1,90ms para o nosso FCP atual.